requirements.txt
Appium-Python-Client>=3.1.0
pytest>=7.4.0
PILLOW>=9.0.0